requirements.txt
# Async HTTP + WebSocket
aiohttp==3.9.5
aiohttp[speedups]          # optional C-accelerated JSON/websocket parsing

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9     # used by Alembic's sync engine for migrations only

# Config / validation
pydantic==2.7.1
pydantic-settings==2.3.1

# Logging
structlog==24.1.0

# CLI reporting
rich==13.7.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
respx==0.21.1
httpx==0.27.0              # required by respx
